requirements.txt
streamlit
sumy
Pillow
